refactor: share code for completion data tables
